fix: address validation vocabulary, workspace contracts, JSONL seal, and source/test boundary
Closes gaps tracked by #1145–#1148 on top of the fleet-coherence README work:
typed validate findings (pass/skip/note/advisory/warning/failure), exclusive
claim and stale-prune regression alignment, sealed legacy JSONL import for
todo rebuild, and release-policy compliance by moving remaining inline tests
under tests/unit/.
