+++ src/decapod/cli.rs
@@ -1809,26 +1809,5 @@ pub(crate) struct DemoCli {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::{BackendType, RepoContext};
-
-    #[test]
-    fn backend_field_selects_the_repository_backend() {
-        let mut context = RepoContext {
-            backend: Some(BackendType::Cloud),
-            ..RepoContext::default()
-        };
-        assert_eq!(context.effective_backend(), BackendType::Cloud);
-
-        context.backend = Some(BackendType::Local);
-        assert_eq!(context.effective_backend(), BackendType::Local);
-    }
-
-    #[test]
-    fn setting_backend_selects_the_canonical_config_field() {
-        let mut context = RepoContext::default();
-        context.set_backend(BackendType::Cloud);
-        assert_eq!(context.backend, Some(BackendType::Cloud));
-        assert_eq!(context.effective_backend(), BackendType::Cloud);
-    }
-}
+#[path = "../../tests/unit/cli_tests.rs"]
+mod tests;
